masters-degree/sirs/backend/requirements.txt
blinker==1.7.0
bson==0.5.10
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
Werkzeug==3.0.1
